Split routes up based on the middleware they require
* Split routes up based on the middleware they require
